requirements.txt
gradio==4.44.0
groq==0.11.0
ollama==0.3.3
requests==2.32.3
python-dotenv==1.0.1
soundfile==0.12.1
numpy==1.26.4
httpx==0.27.2
